untested helio calculations
