+++ src/earth/periodic_terms.rs
@@ -1,0 +1,208 @@
+/// A row from the Earth's Periodic Terms table (from the
+/// NREL SPA document)
+pub struct EarthPeriodicTableRow {
+    pub term : &'static str,
+    pub a : f64,
+    pub b : f64,
+    pub c : f64
+}
+
+/// A collection of Earth's Periodic Terms as per the NREL
+/// SPA document (A.4.2)
+pub const EARTH_PERIODIC_TERMS : &[EarthPeriodicTableRow] = &[
+    EarthPeriodicTableRow { term: "L0", a:    175347046_f64, b:         0_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "L0", a:      3341656_f64, b: 4.6692568_f64, c: 6283.07585_f64 },
+    EarthPeriodicTableRow { term: "L0", a:        34894_f64, b:    4.6261_f64, c: 12566.1517_f64 },
+    EarthPeriodicTableRow { term: "L0", a:         3497_f64, b:    2.7441_f64, c:  5753.3849_f64 },
+    EarthPeriodicTableRow { term: "L0", a:         3418_f64, b:    2.8289_f64, c:     3.5231_f64 },
+    EarthPeriodicTableRow { term: "L0", a:         3136_f64, b:    3.6277_f64, c: 77713.7715_f64 },
+    EarthPeriodicTableRow { term: "L0", a:         2676_f64, b:    4.4181_f64, c:  7860.4194_f64 },
+    EarthPeriodicTableRow { term: "L0", a:         2343_f64, b:    6.1352_f64, c:  3930.2097_f64 },
+    EarthPeriodicTableRow { term: "L0", a:         1324_f64, b:    0.7425_f64, c: 11506.7698_f64 },
+    EarthPeriodicTableRow { term: "L0", a:         1273_f64, b:    2.0371_f64, c:    529.691_f64 },
+    EarthPeriodicTableRow { term: "L0", a:         1199_f64, b:    1.1096_f64, c:  1577.3435_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          990_f64, b:     5.233_f64, c:   5884.927_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          902_f64, b:     2.045_f64, c:     26.298_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          857_f64, b:     3.508_f64, c:    398.149_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          780_f64, b:     1.179_f64, c:   5223.694_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          753_f64, b:     2.533_f64, c:   5507.553_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          505_f64, b:     4.583_f64, c:  18849.228_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          492_f64, b:     4.205_f64, c:    775.523_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          357_f64, b:      2.92_f64, c:      0.067_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          317_f64, b:     5.849_f64, c:  11790.629_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          284_f64, b:     1.899_f64, c:    796.298_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          271_f64, b:     0.315_f64, c:  10977.079_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          243_f64, b:     0.345_f64, c:   5486.778_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          206_f64, b:     4.806_f64, c:   2544.314_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          205_f64, b:     1.869_f64, c:   5573.143_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          202_f64, b:     2.458_f64, c:   6069.777_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          156_f64, b:     0.833_f64, c:    213.299_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          132_f64, b:     3.411_f64, c:   2942.463_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          126_f64, b:     1.083_f64, c:     20.775_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          115_f64, b:     0.645_f64, c:       0.98_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          103_f64, b:     0.636_f64, c:   4694.003_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          102_f64, b:     0.976_f64, c:  15720.839_f64 },
+    EarthPeriodicTableRow { term: "L0", a:          102_f64, b:     4.267_f64, c:      7.114_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           99_f64, b:      6.21_f64, c:    2146.17_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           98_f64, b:      0.68_f64, c:     155.42_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           86_f64, b:      5.98_f64, c:  161000.69_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           85_f64, b:       1.3_f64, c:    6275.96_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           85_f64, b:      3.67_f64, c:    71430.7_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           80_f64, b:      1.81_f64, c:   17260.15_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           79_f64, b:      3.04_f64, c:   12036.46_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           75_f64, b:      1.76_f64, c:    5088.63_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           74_f64, b:       3.5_f64, c:    3154.69_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           74_f64, b:      4.68_f64, c:     801.82_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           70_f64, b:      0.83_f64, c:    9437.76_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           62_f64, b:      3.98_f64, c:    8827.39_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           61_f64, b:      1.82_f64, c:     7084.9_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           57_f64, b:      2.78_f64, c:     6286.6_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           56_f64, b:      4.39_f64, c:    14143.5_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           56_f64, b:      3.47_f64, c:    6279.55_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           52_f64, b:      0.19_f64, c:   12139.55_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           52_f64, b:      1.33_f64, c:    1748.02_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           51_f64, b:      0.28_f64, c:    5856.48_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           49_f64, b:      0.49_f64, c:    1194.45_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           41_f64, b:      5.37_f64, c:    8429.24_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           41_f64, b:       2.4_f64, c:   19651.05_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           39_f64, b:      6.17_f64, c:   10447.39_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           37_f64, b:      6.04_f64, c:   10213.29_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           37_f64, b:      2.57_f64, c:    1059.38_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           36_f64, b:      1.71_f64, c:    2352.87_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           36_f64, b:      1.78_f64, c:    6812.77_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           33_f64, b:      0.59_f64, c:   17789.85_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           30_f64, b:      0.44_f64, c:   83996.85_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           30_f64, b:      2.74_f64, c:    1349.87_f64 },
+    EarthPeriodicTableRow { term: "L0", a:           25_f64, b:      3.16_f64, c:    4690.48_f64 },
+    EarthPeriodicTableRow { term: "L1", a: 628331966747_f64, b:         0_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "L1", a:       206059_f64, b:  2.678235_f64, c: 6283.07585_f64 },
+    EarthPeriodicTableRow { term: "L1", a:         4303_f64, b:    2.6351_f64, c: 12566.1517_f64 },
+    EarthPeriodicTableRow { term: "L1", a:          425_f64, b:      1.59_f64, c:      3.523_f64 },
+    EarthPeriodicTableRow { term: "L1", a:          119_f64, b:     5.796_f64, c:     26.298_f64 },
+    EarthPeriodicTableRow { term: "L1", a:          109_f64, b:     2.966_f64, c:   1577.344_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           93_f64, b:      2.59_f64, c:   18849.23_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           72_f64, b:      1.14_f64, c:     529.69_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           68_f64, b:      1.87_f64, c:     398.15_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           67_f64, b:      4.41_f64, c:    5507.55_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           59_f64, b:      2.89_f64, c:    5223.69_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           56_f64, b:      2.17_f64, c:     155.42_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           45_f64, b:       0.4_f64, c:      796.3_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           36_f64, b:      0.47_f64, c:     775.52_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           29_f64, b:      2.65_f64, c:       7.11_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           21_f64, b:      5.34_f64, c:       0.98_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           19_f64, b:      1.85_f64, c:    5486.78_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           19_f64, b:      4.97_f64, c:      213.3_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           17_f64, b:      2.99_f64, c:    6275.96_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           16_f64, b:      0.03_f64, c:    2544.31_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           16_f64, b:      1.43_f64, c:    2146.17_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           15_f64, b:      1.21_f64, c:   10977.08_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           12_f64, b:      2.83_f64, c:    1748.02_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           12_f64, b:      3.26_f64, c:    5088.63_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           12_f64, b:      5.27_f64, c:    1194.45_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           12_f64, b:      2.08_f64, c:       4694_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           11_f64, b:      0.77_f64, c:     553.57_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           10_f64, b:       1.3_f64, c:     6286.6_f64 },
+    EarthPeriodicTableRow { term: "L1", a:           10_f64, b:      4.24_f64, c:    1349.87_f64 },
+    EarthPeriodicTableRow { term: "L1", a:            9_f64, b:       2.7_f64, c:     242.73_f64 },
+    EarthPeriodicTableRow { term: "L1", a:            9_f64, b:      5.64_f64, c:     951.72_f64 },
+    EarthPeriodicTableRow { term: "L1", a:            8_f64, b:       5.3_f64, c:    2352.87_f64 },
+    EarthPeriodicTableRow { term: "L1", a:            6_f64, b:      2.65_f64, c:    9437.76_f64 },
+    EarthPeriodicTableRow { term: "L1", a:            6_f64, b:      4.67_f64, c:    4690.48_f64 },
+    EarthPeriodicTableRow { term: "L2", a:        52919_f64, b:         0_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "L2", a:         8720_f64, b:    1.0721_f64, c:  6283.0758_f64 },
+    EarthPeriodicTableRow { term: "L2", a:          309_f64, b:     0.867_f64, c:  12566.152_f64 },
+    EarthPeriodicTableRow { term: "L2", a:           27_f64, b:      0.05_f64, c:       3.52_f64 },
+    EarthPeriodicTableRow { term: "L2", a:           16_f64, b:      5.19_f64, c:       26.3_f64 },
+    EarthPeriodicTableRow { term: "L2", a:           16_f64, b:      3.68_f64, c:     155.42_f64 },
+    EarthPeriodicTableRow { term: "L2", a:           10_f64, b:      0.76_f64, c:   18849.23_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            9_f64, b:      2.06_f64, c:   77713.77_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            7_f64, b:      0.83_f64, c:     775.52_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            5_f64, b:      4.66_f64, c:    1577.34_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            4_f64, b:      1.03_f64, c:       7.11_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            4_f64, b:      3.44_f64, c:    5573.14_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            3_f64, b:      5.14_f64, c:      796.3_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            3_f64, b:      6.05_f64, c:    5507.55_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            3_f64, b:      1.19_f64, c:     242.73_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            3_f64, b:      6.12_f64, c:     529.69_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            3_f64, b:      0.31_f64, c:     398.15_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            3_f64, b:      2.28_f64, c:     553.57_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            2_f64, b:      4.38_f64, c:    5223.69_f64 },
+    EarthPeriodicTableRow { term: "L2", a:            2_f64, b:      3.75_f64, c:       0.98_f64 },
+    EarthPeriodicTableRow { term: "L3", a:          289_f64, b:     5.844_f64, c:   6283.076_f64 },
+    EarthPeriodicTableRow { term: "L3", a:           35_f64, b:         0_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "L3", a:           17_f64, b:      5.49_f64, c:   12566.15_f64 },
+    EarthPeriodicTableRow { term: "L3", a:            3_f64, b:       5.2_f64, c:     155.42_f64 },
+    EarthPeriodicTableRow { term: "L3", a:            1_f64, b:      4.72_f64, c:       3.52_f64 },
+    EarthPeriodicTableRow { term: "L3", a:            1_f64, b:       5.3_f64, c:   18849.23_f64 },
+    EarthPeriodicTableRow { term: "L3", a:            1_f64, b:      5.97_f64, c:     242.73_f64 },
+    EarthPeriodicTableRow { term: "L4", a:          114_f64, b:     3.142_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "L4", a:            8_f64, b:      4.13_f64, c:    6283.08_f64 },
+    EarthPeriodicTableRow { term: "L4", a:            1_f64, b:      3.84_f64, c:   12566.15_f64 },
+    EarthPeriodicTableRow { term: "L5", a:            1_f64, b:      3.14_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "B0", a:          280_f64, b:     3.199_f64, c:  84334.662_f64 },
+    EarthPeriodicTableRow { term: "B0", a:          102_f64, b:     5.422_f64, c:   5507.553_f64 },
+    EarthPeriodicTableRow { term: "B0", a:           80_f64, b:      3.88_f64, c:    5223.69_f64 },
+    EarthPeriodicTableRow { term: "B0", a:           44_f64, b:       3.7_f64, c:    2352.87_f64 },
+    EarthPeriodicTableRow { term: "B0", a:           32_f64, b:         4_f64, c:    1577.34_f64 },
+    EarthPeriodicTableRow { term: "B1", a:            9_f64, b:       3.9_f64, c:    5507.55_f64 },
+    EarthPeriodicTableRow { term: "B1", a:            6_f64, b:      1.73_f64, c:    5223.69_f64 },
+    EarthPeriodicTableRow { term: "R0", a:    100013989_f64, b:         0_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "R0", a:      1670700_f64, b: 3.0984635_f64, c: 6283.07585_f64 },
+    EarthPeriodicTableRow { term: "R0", a:        13956_f64, b:   3.05525_f64, c: 12566.1517_f64 },
+    EarthPeriodicTableRow { term: "R0", a:         3084_f64, b:    5.1985_f64, c: 77713.7715_f64 },
+    EarthPeriodicTableRow { term: "R0", a:         1628_f64, b:    1.1739_f64, c:  5753.3849_f64 },
+    EarthPeriodicTableRow { term: "R0", a:         1576_f64, b:    2.8469_f64, c:  7860.4194_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          925_f64, b:     5.453_f64, c:   11506.77_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          542_f64, b:     4.564_f64, c:    3930.21_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          472_f64, b:     3.661_f64, c:   5884.927_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          346_f64, b:     0.964_f64, c:   5507.553_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          329_f64, b:       5.9_f64, c:   5223.694_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          307_f64, b:     0.299_f64, c:   5573.143_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          243_f64, b:     4.273_f64, c:  11790.629_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          212_f64, b:     5.847_f64, c:   1577.344_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          186_f64, b:     5.022_f64, c:  10977.079_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          175_f64, b:     3.012_f64, c:  18849.228_f64 },
+    EarthPeriodicTableRow { term: "R0", a:          110_f64, b:     5.055_f64, c:   5486.778_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           98_f64, b:      0.89_f64, c:    6069.78_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           86_f64, b:      5.69_f64, c:   15720.84_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           86_f64, b:      1.27_f64, c:  161000.69_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           65_f64, b:      0.27_f64, c:   17260.15_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           63_f64, b:      0.92_f64, c:     529.69_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           57_f64, b:      2.01_f64, c:   83996.85_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           56_f64, b:      5.24_f64, c:    71430.7_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           49_f64, b:      3.25_f64, c:    2544.31_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           47_f64, b:      2.58_f64, c:     775.52_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           45_f64, b:      5.54_f64, c:    9437.76_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           43_f64, b:      6.01_f64, c:    6275.96_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           39_f64, b:      5.36_f64, c:       4694_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           38_f64, b:      2.39_f64, c:    8827.39_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           37_f64, b:      0.83_f64, c:   19651.05_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           37_f64, b:       4.9_f64, c:   12139.55_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           36_f64, b:      1.67_f64, c:   12036.46_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           35_f64, b:      1.84_f64, c:    2942.46_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           33_f64, b:      0.24_f64, c:     7084.9_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           32_f64, b:      0.18_f64, c:    5088.63_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           32_f64, b:      1.78_f64, c:     398.15_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           28_f64, b:      1.21_f64, c:     6286.6_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           28_f64, b:       1.9_f64, c:    6279.55_f64 },
+    EarthPeriodicTableRow { term: "R0", a:           26_f64, b:      4.59_f64, c:   10447.39_f64 },
+    EarthPeriodicTableRow { term: "R1", a:       103019_f64, b:   1.10749_f64, c: 6283.07585_f64 },
+    EarthPeriodicTableRow { term: "R1", a:         1721_f64, b:    1.0644_f64, c: 12566.1517_f64 },
+    EarthPeriodicTableRow { term: "R1", a:          702_f64, b:     3.142_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "R1", a:           32_f64, b:      1.02_f64, c:   18849.23_f64 },
+    EarthPeriodicTableRow { term: "R1", a:           31_f64, b:      2.84_f64, c:    5507.55_f64 },
+    EarthPeriodicTableRow { term: "R1", a:           25_f64, b:      1.32_f64, c:    5223.69_f64 },
+    EarthPeriodicTableRow { term: "R1", a:           18_f64, b:      1.42_f64, c:    1577.34_f64 },
+    EarthPeriodicTableRow { term: "R1", a:           10_f64, b:      5.91_f64, c:   10977.08_f64 },
+    EarthPeriodicTableRow { term: "R1", a:            9_f64, b:      1.42_f64, c:    6275.96_f64 },
+    EarthPeriodicTableRow { term: "R1", a:            9_f64, b:      0.27_f64, c:    5486.78_f64 },
+    EarthPeriodicTableRow { term: "R2", a:         4359_f64, b:    5.7846_f64, c:  6283.0758_f64 },
+    EarthPeriodicTableRow { term: "R2", a:          124_f64, b:     5.579_f64, c:  12566.152_f64 },
+    EarthPeriodicTableRow { term: "R2", a:           12_f64, b:      3.14_f64, c:          0_f64 },
+    EarthPeriodicTableRow { term: "R2", a:            9_f64, b:      3.63_f64, c:   77713.77_f64 },
+    EarthPeriodicTableRow { term: "R2", a:            6_f64, b:      1.87_f64, c:    5573.14_f64 },
+    EarthPeriodicTableRow { term: "R2", a:            3_f64, b:      5.47_f64, c:   18849.23_f64 },
+    EarthPeriodicTableRow { term: "R3", a:          145_f64, b:     4.273_f64, c:   6283.076_f64 },
+    EarthPeriodicTableRow { term: "R3", a:            7_f64, b:      3.92_f64, c:   12566.15_f64 },
+    EarthPeriodicTableRow { term: "R4", a:            4_f64, b:      2.56_f64, c:    6283.08_f64 } 
+];
